Move things around a bit into an EmailService (Utils really needs to go away at some point)
