+++ OurUmbraco.Site/Views/Partials/Emails/ActivationEmailTemplate.cshtml
@@ -1,4 +1,4 @@
-﻿@inherits UmbracoViewPage<OurUmbraco.Emails.Models.ActivationEmailModel>
+﻿@inherits UmbracoViewPage<OurUmbraco.Our.Models.ActivationEmailModel>
 
 @{
     Layout = "EmailTemplateMaster.cshtml";
