+++ OurUmbraco/Forum/Library/Utils.cs
@@ -1,16 +1,11 @@
 ﻿using System;
-using System.Collections.Specialized;
 using System.Configuration;
 using System.Linq;
-using System.Net;
 using System.Net.Mail;
 using System.Web;
-using OurUmbraco.Emails;
-using OurUmbraco.Emails.Models;
 using OurUmbraco.Forum.Extensions;
 using umbraco.BusinessLogic;
 using Umbraco.Core;
-using Umbraco.Core.Logging;
 using Umbraco.Core.Models;
 using Umbraco.Web;
 
@@ -107,61 +102,6 @@ namespace OurUmbraco.Forum.Library
             {
                 Log.Add(LogTypes.Error, new User(0), -1, string.Format("Error sending potential spam member notification: {0} {1} {2}",
                     ex.Message, ex.StackTrace, ex.InnerException));
-            }
-        }
-
-        public static void SendActivationMail(IMember member)
-        {
-            try
-            {
-                const string subject = "Activate your account on our.umbraco.org";
-                
-                var body = EmailsUtils.Mvc.RenderPartial("Partials/Emails/ActivationEmailTemplate.cshtml", 
-                    new ActivationEmailModel(member));
-
-                var mailMessage = new MailMessage
-                {
-                    Subject = subject,
-                    Body = body.ToString(),
-                    IsBodyHtml = true
-                };
-                
-                mailMessage.To.Add(member.Email);
-
-                mailMessage.From = new MailAddress("robot@umbraco.org");
-
-                var smtpClient = new SmtpClient();
-                smtpClient.Send(mailMessage);
-            }
-            catch (Exception ex)
-            {
-                var error = string.Format("*ERROR* sending activation mail for member {0} - {1} {2} {3}", member.Email, ex.Message, ex.StackTrace, ex.InnerException);
-                SendSlackNotification(error);
-                LogHelper.Error<Utils>(error, ex);
-            }
-        }
-
-        private static void SendSlackNotification(string body)
-        {
-            using (var client = new WebClient())
-            {
-                var values = new NameValueCollection
-                {
-                    {"channel", ConfigurationManager.AppSettings["SlackChannel"]},
-                    {"token", ConfigurationManager.AppSettings["SlackToken"]},
-                    {"username", ConfigurationManager.AppSettings["SlackUsername"]},
-                    {"icon_url", ConfigurationManager.AppSettings["SlackIconUrl"]},
-                    {"text", body}
-                };
-
-                try
-                {
-                    client.UploadValues("https://slack.com/api/chat.postMessage", "POST", values);
-                }
-                catch (Exception ex)
-                {
-                    LogHelper.Error<Utils>("Posting update to Slack failed", ex);
-                }
             }
         }
